// types.ts
export type ApiProvider = "geminiDefault" | "geminiCustom";